@@ -3,13 +3,23 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start"
+    "start": "node node_modules/react-native/local-cli/cli.js start",
+    "test": "jest"
   },
   "dependencies": {
-    "react": "^0.14.8",
-    "react-native": "^0.25.1",
-    "react-native-button": "^1.5.0",
-    "react-native-modalbox": "^1.3.8",
-    "react-native-slider": "^0.4.0"
+    "react": "16.0.0-alpha.12",
+    "react-native": "0.45.1",
+    "react-native-button": "^2.0.0",
+    "react-native-modalbox": "^1.4.0",
+    "react-native-slider": "^0.10.0"
+  },
+  "devDependencies": {
+    "babel-jest": "20.0.3",
+    "babel-preset-react-native": "2.0.0",
+    "jest": "20.0.4",
+    "react-test-renderer": "16.0.0-alpha.12"
+  },
+  "jest": {
+    "preset": "react-native"
   }
 }
