@@ -1,20 +1,116 @@
-'use strict';
+import React from 'react';
+import Button from 'react-native-button';
+import Modal from 'react-native-modalbox';
+import Slider from 'react-native-slider';
 
-var React   = require('react');
-var Button  = require('react-native-button');
-var Modal   = require('react-native-modalbox');
-var Slider  = require('react-native-slider');
-var window  = require('Dimensions').get('window');
-
-var {
+import {
   AppRegistry,
-  StyleSheet,
   Text,
+  StyleSheet,
+  ScrollView,
   View,
-  ScrollView
-} = require('react-native');
+  Dimensions,
+  TextInput
+} from 'react-native';
 
-var styles = StyleSheet.create({
+var screen = Dimensions.get('window');
+
+class Example extends React.Component {
+
+  constructor() {
+    super();
+    this.state = {
+      isOpen: false,
+      isDisabled: false,
+      swipeToClose: true,
+      sliderValue: 0.3
+    };
+  }
+
+  onClose() {
+    console.log('Modal just closed');
+  }
+
+  onOpen() {
+    console.log('Modal just openned');
+  }
+
+  onClosingState(state) {
+    console.log('the open/close of the swipeToClose just changed');
+  }
+
+  renderList() {
+    var list = [];
+
+    for (var i=0;i<50;i++) {
+      list.push(<Text style={styles.text} key={i}>Elem {i}</Text>);
+    }
+
+    return list;
+  }
+
+  render() {
+    var BContent = <Button onPress={() => this.setState({isOpen: false})} style={[styles.btn, styles.btnModal]}>X</Button>;
+
+    return (
+      <View style={styles.wrapper}>
+        <Button onPress={() => this.refs.modal1.open()} style={styles.btn}>Basic modal</Button>
+        <Button onPress={() => this.refs.modal2.open()} style={styles.btn}>Position top</Button>
+        <Button onPress={() => this.refs.modal3.open()} style={styles.btn}>Position centered + backdrop + disable</Button>
+        <Button onPress={() => this.refs.modal4.open()} style={styles.btn}>Position bottom + backdrop + slider</Button>
+        <Button onPress={() => this.setState({isOpen: true})} style={styles.btn}>Backdrop + backdropContent</Button>
+        <Button onPress={() => this.refs.modal6.open()} style={styles.btn}>Position bottom + ScrollView</Button>
+        <Button onPress={() => this.refs.modal7.open()} style={styles.btn}>Modal with keyboard support</Button>
+
+        <Modal
+          style={[styles.modal, styles.modal1]}
+          ref={"modal1"}
+          swipeToClose={this.state.swipeToClose}
+          onClosed={this.onClose}
+          onOpened={this.onOpen}
+          onClosingState={this.onClosingState}>
+            <Text style={styles.text}>Basic modal</Text>
+            <Button onPress={() => this.setState({swipeToClose: !this.state.swipeToClose})} style={styles.btn}>Disable swipeToClose({this.state.swipeToClose ? "true" : "false"})</Button>
+        </Modal>
+
+        <Modal style={[styles.modal, styles.modal2]} backdrop={false}  position={"top"} ref={"modal2"}>
+          <Text style={[styles.text, {color: "white"}]}>Modal on top</Text>
+        </Modal>
+
+        <Modal style={[styles.modal, styles.modal3]} position={"center"} ref={"modal3"} isDisabled={this.state.isDisabled}>
+          <Text style={styles.text}>Modal centered</Text>
+          <Button onPress={() => this.setState({isDisabled: !this.state.isDisabled})} style={styles.btn}>Disable ({this.state.isDisabled ? "true" : "false"})</Button>
+        </Modal>
+
+        <Modal style={[styles.modal, styles.modal4]} position={"bottom"} ref={"modal4"}>
+          <Text style={styles.text}>Modal on bottom with backdrop</Text>
+          <Slider style={{width: 200}} value={this.state.sliderValue} onValueChange={(value) => this.setState({sliderValue: value})} />
+        </Modal>
+
+        <Modal isOpen={this.state.isOpen} onClosed={() => this.setState({isOpen: false})} style={[styles.modal, styles.modal4]} position={"center"} backdropContent={BContent}>
+          <Text style={styles.text}>Modal with backdrop content</Text>
+        </Modal>
+
+        <Modal style={[styles.modal, styles.modal4]} position={"bottom"} ref={"modal6"} swipeArea={20}>
+          <ScrollView>
+            <View style={{width: screen.width, paddingLeft: 10}}>
+              {this.renderList()}
+            </View>
+          </ScrollView>
+        </Modal>
+
+        <Modal ref={"modal7"} style={[styles.modal, styles.modal4]} position={"center"}>
+          <View>
+            <TextInput style={{height: 50, width: 200, backgroundColor: '#DDDDDD'}}/>
+          </View>
+        </Modal>
+      </View>
+    );
+  }
+
+}
+
+const styles = StyleSheet.create({
 
   wrapper: {
     paddingTop: 50,
@@ -61,123 +157,6 @@ var styles = StyleSheet.create({
     fontSize: 22
   }
 
-});
-
-var Example = React.createClass({
-
-  getInitialState: function() {
-    return {
-      isOpen: false,
-      isDisabled: false,
-      swipeToClose: true,
-      sliderValue: 0.3
-    };
-  },
-
-  openModal1: function(id) {
-    this.refs.modal1.open();
-  },
-
-  openModal2: function(id) {
-    this.refs.modal2.open();
-  },
-
-  openModal3: function(id) {
-    this.refs.modal3.open();
-  },
-
-  openModal4: function(id) {
-    this.refs.modal4.open();
-  },
-
-  openModal5: function(id) {
-    this.setState({isOpen: true});
-  },
-
-  closeModal5: function(id) {
-    this.setState({isOpen: false});
-  },
-
-  openModal6: function(id) {
-    this.refs.modal6.open();
-  },
-
-  toggleDisable: function() {
-    this.setState({isDisabled: !this.state.isDisabled});
-  },
-
-  toggleSwipeToClose: function() {
-    this.setState({swipeToClose: !this.state.swipeToClose});
-  },
-
-  onClose: function() {
-    console.log('Modal just closed');
-  },
-
-  onOpen: function() {
-    console.log('Modal just openned');
-  },
-
-  onClosingState: function(state) {
-    console.log('the open/close of the swipeToClose just changed');
-  },
-
-  renderList() {
-    var list = [];
-
-    for (var i=0;i<50;i++) {
-      list.push(<Text style={styles.text} key={i}>Elem {i}</Text>);
-    }
-
-    return list;
-  },
-
-  render: function() {
-
-    var BContent = <Button onPress={this.closeModal5} style={[styles.btn, styles.btnModal]}>X</Button>;
-
-    return (
-      <View style={styles.wrapper}>
-        <Button onPress={this.openModal1} style={styles.btn}>Basic modal</Button>
-        <Button onPress={this.openModal2} style={styles.btn}>Position top</Button>
-        <Button onPress={this.openModal3} style={styles.btn}>Position centered + backdrop + disable</Button>
-        <Button onPress={this.openModal4} style={styles.btn}>Position bottom + backdrop + slider</Button>
-        <Button onPress={this.openModal5} style={styles.btn}>Backdrop + backdropContent</Button>
-        <Button onPress={this.openModal6} style={styles.btn}>Position bottom + ScrollView</Button>
-
-        <Modal style={[styles.modal, styles.modal1]} ref={"modal1"} swipeToClose={this.state.swipeToClose} onClosed={this.onClose} onOpened={this.onOpen} onClosingState={this.onClosingState}>
-          <Text style={styles.text}>Basic modal</Text>
-          <Button onPress={this.toggleSwipeToClose} style={styles.btn}>Disable swipeToClose({this.state.swipeToClose ? "true" : "false"})</Button>
-        </Modal>
-
-        <Modal style={[styles.modal, styles.modal2]} backdrop={false}  position={"top"} ref={"modal2"}>
-          <Text style={[styles.text, {color: "white"}]}>Modal on top</Text>
-        </Modal>
-
-        <Modal style={[styles.modal, styles.modal3]} position={"center"} ref={"modal3"} isDisabled={this.state.isDisabled}>
-          <Text style={styles.text}>Modal centered</Text>
-          <Button onPress={this.toggleDisable} style={styles.btn}>Disable ({this.state.isDisabled ? "true" : "false"})</Button>
-        </Modal>
-
-        <Modal style={[styles.modal, styles.modal4]} position={"bottom"} ref={"modal4"}>
-          <Text style={styles.text}>Modal on bottom with backdrop</Text>
-          <Slider style={{width: 200}} value={this.state.sliderValue} onValueChange={(value) => this.setState({sliderValue: value})} />
-        </Modal>
-
-        <Modal isOpen={this.state.isOpen} onClosed={this.closeModal5} style={[styles.modal, styles.modal4]} position={"center"} backdropContent={BContent}>
-          <Text style={styles.text}>Modal with backdrop content</Text>
-        </Modal>
-
-        <Modal style={[styles.modal, styles.modal4]} position={"bottom"} ref={"modal6"} swipeArea={20}>
-          <ScrollView>
-            <View style={{width: window.width, paddingLeft: 10}}>
-              {this.renderList()}
-            </View>
-          </ScrollView>
-        </Modal>
-      </View>
-    );
-  }
 });
 
 AppRegistry.registerComponent('Example', () => Example);
